Fix resizing bug
If an instance is being moved to another host during resizing, instance
root disk is not mapped to the new host, but is mapped to the old host.
In this case the disk must not be directly unmapped from the new host,
but must be deleted with ignore_mappings flag.

Also since 12.0.4 driver.py checks the disks for existence before
removing its snapshot. Because that this check is stubbed with False to
enable create_image call from cache method, the snapshot is not removed.
To workaround this create_snapshot now creates the snapshot even the
snapshot is already exists. It does not avoid SIO from unused snapshot,
but allows the second resize at least.

Also extend operation is now called even if size is not changed. This
allows to resize an instance to a flavor with the same size, but
different storage pool/protection domain settings.