Fix bug of snapshot handling
A confusing fix was merged to Nova, which checks if an image exists
before to delete its snapshot. This is a workaround to rbd utils feature
which fails if the image is absent. The right way to fix this bug is to
handle this case in rbd utils itself.

check_image_exists method can not check if an image exists indeed, but
always returns False (see comments in the method). Because this and
above all right ways to fix this bug are very complex. So that an empty
dumb workaround is added in the fix here.